+++ src/main/scala/leapfrogTriejoin/LeapfrogJoinInterface.scala
@@ -4,7 +4,7 @@ abstract class LeapfrogJoinInterface {
 
   def init(): Unit
   def leapfrogNext(): Unit
-  def leapfrogSeek(key: Long): Unit
-  def key: Long
+  def leapfrogSeek(key: Int): Unit
+  def key: Int
   def atEnd: Boolean
 }
